Alternate fix to mypy errors

+++ packages/syft/tests/syft/core/tensor/smpc/smpc_tensor_test.py
@@ -135,7 +135,7 @@ def test_mpc_forward_methods(method_str: str, kwargs: TypeDict[str, Any]) -> Non
 
 
 def test_repeat() -> None:
-    value = np.array([[1, 2], [3, 4]])
+    value = np.array([[1, 2], [3, 4]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -150,7 +150,7 @@ def test_repeat() -> None:
 
 
 def test_copy() -> None:
-    value = np.array([[1, 2], [3, 4]])
+    value = np.array([[1, 2], [3, 4]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -166,7 +166,7 @@ def test_copy() -> None:
 
 
 def test_diagonal() -> None:
-    value = np.array([[0, 1], [2, 3]])
+    value = np.array([[0, 1], [2, 3]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -181,7 +181,7 @@ def test_diagonal() -> None:
 
 
 def test_flatten() -> None:
-    value = np.array([[89, 12, 54], [412, 89, 42], [87, 32, 58]])
+    value = np.array([[89, 12, 54], [412, 89, 42], [87, 32, 58]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -196,7 +196,7 @@ def test_flatten() -> None:
 
 
 def test_transpose() -> None:
-    value = np.array([[89, 12, 54], [412, 89, 42], [87, 32, 58]])
+    value = np.array([[89, 12, 54], [412, 89, 42], [87, 32, 58]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -211,7 +211,7 @@ def test_transpose() -> None:
 
 
 def test_resize() -> None:
-    value = np.array([[89, 12], [412, 89], [87, 32]])
+    value = np.array([[89, 12], [412, 89], [87, 32]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -227,7 +227,7 @@ def test_resize() -> None:
 
 
 def test_ravel() -> None:
-    value = np.array([[8, 1, 5], [4, 8, 4], [7, 2, 27]])
+    value = np.array([[8, 1, 5], [4, 8, 4], [7, 2, 27]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -242,7 +242,7 @@ def test_ravel() -> None:
 
 
 def test_compress() -> None:
-    value = np.array([[1, 2], [3, 4], [5, 6]])
+    value = np.array([[1, 2], [3, 4], [5, 6]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -257,7 +257,7 @@ def test_compress() -> None:
 
 
 def test_reshape() -> None:
-    value = np.array([[1, 2], [3, 4], [5, 6]])
+    value = np.array([[1, 2], [3, 4], [5, 6]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -272,7 +272,7 @@ def test_reshape() -> None:
 
 
 def test_squeeze() -> None:
-    value = np.array([[7], [6], [72]])
+    value = np.array([[7], [6], [72]]).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
 
@@ -287,7 +287,7 @@ def test_squeeze() -> None:
 
 
 def test_swapaxes() -> None:
-    value = np.array(np.array([[613, 645, 738], [531, 412, 658]]))
+    value = np.array(np.array([[613, 645, 738], [531, 412, 658]])).astype(np.int32)
 
     remote_value = clients[0].syft.core.tensor.tensor.Tensor(value)
